Move PianoSpacerProtocol to new file
